abstracted logs query parsing and preparation
